Move __main__ demos to examples
