+++ interactive_knowledge_importer.py
@@ -150,9 +150,4 @@ def run_interactive_importer():
         else:
             logger.warning(f"Не удалось подготовить документы для индексации для QID {selected_qid}.")
 
-if __name__ == "__main__":
-    # Перед запуском этого скрипта, убедись, что LlamaIndex Settings 
-    # (LLM и EmbedModel) корректно инициализируются при импорте 
-    # semantic_memory_index.py (что обычно происходит, если они там в глобальной области видимости).
-    # Также убедись, что папка для индекса существует или может быть создана.
-    run_interactive_importer()
+
