+++ sris_kernel.py
@@ -408,11 +408,4 @@ class SRISChatApp:
         thread.daemon = True 
         thread.start()
 
-if __name__ == "__main__":
-    print("--- SRIS Kernel с GUI ---")
-    print("--- Логи работы ядра будут выводиться в терминале. ---")
-    print("--- Закройте окно GUI для завершения программы. ---")
-    main_window = tk.Tk()
-    app = SRISChatApp(main_window)
-    main_window.mainloop()
-    print("\n--- Программа SRIS завершена ---")
+
